Split utils module
