+++ sundog/management/commands/overwrite_document_template.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 from sundog.models import Document
-from sundog.utils import mutate
+from sundog.util.functional import mutate
 
 
 class Command(BaseCommand):
